requirements.txt
streamlit>=1.40.0
supabase>=2.0.0
httpx>=0.27.0
plotly>=5.18.0
python-dotenv>=1.0.0
bcrypt>=4.0.0
pandas>=2.0.0
apscheduler>=3.10.0
